Move transaction, atomic_transaction to sqlite.connection
